change some for statement to c++11 range-based for

+++ Lib/Node/Node.cpp
@@ -2,7 +2,7 @@
 
 int Node::searchConnectNode( const string &name )
 {
-	for( register int i = 0 ; i < static_cast<int>( m_connect.size() ) ; i++ )
+	for( register unsigned int i = 0 ; i < m_connect.size() ; i++ )
 		 if( m_connect[i]->name() == name ) return i;
 	return -1;
 }
